Fix: handle timeframe, remove duplicated stuff
