+++ front/lib/actions/mcp_internal_actions/servers/child_agent_debugger.ts
@@ -1,11 +1,11 @@
-import {
-  CHILD_AGENT_CONFIGURATION_URI_PATTERN,
-  INTERNAL_MIME_TYPES,
-} from "@dust-tt/client";
+import { INTERNAL_MIME_TYPES } from "@dust-tt/client";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { z } from "zod";
 
-import { ConfigurableToolInputSchemas } from "@app/lib/actions/mcp_internal_actions/input_schemas";
+import {
+  CHILD_AGENT_CONFIGURATION_URI_PATTERN,
+  ConfigurableToolInputSchemas,
+} from "@app/lib/actions/mcp_internal_actions/input_schemas";
 import type { InternalMCPServerDefinitionType } from "@app/lib/api/mcp";
 import type { Result } from "@app/types";
 import { Err, Ok } from "@app/types";
