+++ front/package.json
@@ -132,6 +132,7 @@
     "uuid": "^9.0.0",
     "yargs": "^17.7.2",
     "zod": "^3.23.8",
+    "zod-to-json-schema": "^3.24.5",
     "zod-validation-error": "^3.4.0"
   },
   "devDependencies": {
