+++ sdks/js/package.json
@@ -44,8 +44,7 @@
     "axios": "^1.7.9",
     "eventsource-parser": "^1.1.1",
     "moment-timezone": "^0.5.46",
-    "zod": "^3.23.8",
-    "zod-to-json-schema": "^3.24.5"
+    "zod": "^3.23.8"
   },
   "browser": {
     "tls": false,
