+++ sdks/js/src/index.ts
@@ -83,7 +83,6 @@ import {
 } from "./types";
 
 export * from "./internal_mime_types";
-export * from "./tool_input_schemas";
 export * from "./types";
 
 interface DustResponse {
