+++ sdks/js/src/internal_mime_types.ts
@@ -216,7 +216,7 @@ const TOOL_MIME_TYPES = {
       "BOOLEAN",
       "ENUM",
       "REASONING_MODEL",
-      "TIME_FRAME",
+      "NULLABLE_TIME_FRAME",
     ],
   }),
   TOOL_OUTPUT: generateToolMimeTypes({
